requirements.txt
annotated-types==0.7.0
caldav==2.2.6
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
greenlet==3.3.2
h11==0.16.0
icalendar==7.0.2
icalendar-searcher==1.0.5
jh2==5.0.10
lxml==6.0.2
niquests==3.17.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
qh3==1.5.6
recurring-ical-events==3.8.1
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.47
sqlmodel==0.0.37
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3-future==2.16.900
wassima==2.0.5
x-wr-timezone==2.0.1
